src/module_dashboard_review_backend/requirements.txt
# Dashboard Review Backend - Requirements
# La Máquina de Noticias - MVP Fase 1
# Backend API para revisión editorial

# Core Framework
fastapi>=0.115.12
uvicorn[standard]>=0.30.0

# Database
supabase-py>=2.4.0

# Data Validation & Settings
pydantic>=2.10.0
pydantic-settings>=2.0.0

# Environment Configuration
python-dotenv>=1.0.0

# Logging
loguru>=0.7.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
httpx>=0.27.0
